refactor(engine): Adhere to CodeReview points
- Move HistoryCleanupConfiguration under org.camunda.bpm.engine.impl.jobexecutor.historycleanup package

Related to: #2728
